+++ src/app/material-timepicker/ngx-material-timepicker.component.ts
@@ -74,7 +74,9 @@ export class NgxMaterialTimepickerComponent implements OnInit, OnDestroy {
     }
 
     @Output() timeSet = new EventEmitter<string>();
+    @Output() opened = new EventEmitter<null>();
     @Output() closed = new EventEmitter<null>();
+    @Output() hourSelected = new EventEmitter<number>();
 
     @ViewChild('timepickerww') timepickerComponent: ElementRef;
 
@@ -133,6 +135,11 @@ export class NgxMaterialTimepickerComponent implements OnInit, OnDestroy {
         this.timepickerService.hour = hour;
     }
 
+    onHourSelected(hour: number): void {
+        this.changeTimeUnit(TimeUnit.MINUTE);
+        this.hourSelected.next(hour);
+    }
+
     onMinuteChange(minute: ClockFaceTime): void {
         this.timepickerService.minute = minute;
     }
@@ -141,11 +148,11 @@ export class NgxMaterialTimepickerComponent implements OnInit, OnDestroy {
         this.timepickerService.period = period;
     }
 
-    changeTimeUnit(unit: TimeUnit) {
+    changeTimeUnit(unit: TimeUnit): void {
         this.activeTimeUnit = unit;
     }
 
-    setTime() {
+    setTime(): void {
         this.timeSet.next(this.timepickerService.getFullTime(this.format));
         this.close();
     }
@@ -155,12 +162,13 @@ export class NgxMaterialTimepickerComponent implements OnInit, OnDestroy {
             time, this.minTime as Moment, this.maxTime as Moment, this.format, this.minutesGap);
     }
 
-    open() {
+    open(): void {
         this.isOpened = true;
         this.animationState = AnimationState.ENTER;
+        this.opened.next();
     }
 
-    close() {
+    close(): void {
         this.animationState = AnimationState.LEAVE;
     }